move httpbinding factory to mission editor
